+++ src/unpoly-migrate/viewport.js
@@ -5,6 +5,9 @@
 up.migrate.renamedPackage('layout', 'viewport')
 
 up.migrate.renamedProperty(up.viewport.config, 'viewports', 'viewportSelectors')
+up.migrate.renamedProperty(up.viewport.config, 'fixedTop', 'fixedTopSelectors')
+up.migrate.renamedProperty(up.viewport.config, 'fixedBottom', 'fixedBottomSelectors')
+up.migrate.renamedProperty(up.viewport.config, 'anchoredRight', 'anchoredRightSelectors')
 up.migrate.renamedProperty(up.viewport.config, 'snap', 'revealSnap')
 up.migrate.removedProperty(up.viewport.config, 'scrollSpeed')